# swagger-generated/go-client/model_address.go
/*
 * Order Service API
 *
 * API for managing orders in the e-commerce platform
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Address struct {
	Street string `json:"street"`
	City string `json:"city"`
	State string `json:"state"`
	PostalCode string `json:"postalCode"`
	Country string `json:"country"`
}