# swagger-generated/go-client/model_order_create.go
/*
 * Order Service API
 *
 * API for managing orders in the e-commerce platform
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type OrderCreate struct {
	UserId string `json:"userId"`
	Items []OrderItem `json:"items"`
	ShippingAddress *Address `json:"shippingAddress"`
	BillingAddress *Address `json:"billingAddress"`
}