# swagger-generated/go-client/model_order_item.go
/*
 * Order Service API
 *
 * API for managing orders in the e-commerce platform
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type OrderItem struct {
	ProductId string `json:"productId"`
	Quantity int32 `json:"quantity"`
	UnitPrice float64 `json:"unitPrice"`
}