# swagger-generated/go-client/model_order_status.go
/*
 * Order Service API
 *
 * API for managing orders in the e-commerce platform
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type OrderStatus string

// List of OrderStatus
const (
	CREATED_OrderStatus OrderStatus = "CREATED"
	PAID_OrderStatus OrderStatus = "PAID"
	SHIPPED_OrderStatus OrderStatus = "SHIPPED"
	DELIVERED_OrderStatus OrderStatus = "DELIVERED"
	CANCELLED_OrderStatus OrderStatus = "CANCELLED"
)