# swagger-generated/go-server/go/model_order_status.go
/*
 * Order Service API
 *
 * API for managing orders in the e-commerce platform
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type OrderStatus string

// List of OrderStatus
const (
	CREATED OrderStatus = "CREATED"
	PAID OrderStatus = "PAID"
	SHIPPED OrderStatus = "SHIPPED"
	DELIVERED OrderStatus = "DELIVERED"
	CANCELLED OrderStatus = "CANCELLED"
)